Issue #73 review comments on renaming CursorImpl and moving to its own class
